programmerDto: add class dto

+++ src/main/java/com/project/app/dto/ProgramerDTO.java
@@ -2,7 +2,7 @@ package com.project.app.dto;
 
 public class ProgramerDTO extends HumanDTO {
 
-    public ProgramerDTO(int id, int age, String sex, double salary, String type) {
+    public ProgramerDTO(Long id, int age, String sex, double salary, String type) {
         super(id, age, sex, salary, type);
     }
